refactor: move controllers to the right package
This time it's correct...
